feat: finished all problem

+++ exercises/23_conversions/using_as.rs
@@ -10,16 +10,22 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
     let values = [3.5, 0.3, 13.0, 11.7];
     println!("{}", average(&values));
+
+
+    // // ini sering
+    // let index = 0i32;
+    //
+    // let arr = vec![1, 2, 3, 4, 5];
+    // let value = arr[index as usize];
 }
 
 #[cfg(test)]
